Chatroom added
The chatroom works, but there is still a lot to do. Serverlogs are still to much, code is not well commented and messages are send as a string. They could be sent as an object with the username and the message as an attribute - in this case it could be a server task not to send the message back to the sender. There are probably a lot performance improvements to make.